third_party/WebKit/LayoutTests/bluetooth/getCharacteristics/service-is-removed.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  testRunner.setBluetoothMockDataSet('HeartRateAdapter');
  return requestDeviceWithKeyDown({filters: [{services: ['heart_rate']}]})
    .then(device => device.gatt.connect())
    .then(gattService => gattService.getPrimaryService('heart_rate'))
    .then(service => {
      testRunner.setBluetoothMockDataSet('MissingServiceHeartRateAdapter');
      return assert_promise_rejects_with_message(
        service.getCharacteristics(),
        new DOMException('GATT Service no longer exists.',
                         'InvalidStateError'),
        'Service got removed.');
    });
}, 'Service is removed. Reject with InvalidStateError.');
</script>
